push

+++ Project_CK/project_ck/src/main/java/tin/ta/project_ck/Repo/SinhVienRepo.java
@@ -1,8 +1,10 @@
 package tin.ta.project_ck.Repo;
-import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import tin.ta.project_ck.Models.StudentModel;
+
 @Repository
-public interface SinhVienRepo extends JpaRepository<StudentModel, Integer>  {
+public interface SinhVienRepo extends JpaRepository<StudentModel, Long> {
+    // Các phương thức truy vấn tùy chỉnh có thể thêm vào đây nếu cần
 }
